+++ packages/blade/src/components/Card/Card.tsx
@@ -96,12 +96,15 @@ export type CardProps = {
    */
   minWidth?: BoxProps['minWidth'];
   /**
+   * Sets maximum width of the card
+   */
+  maxWidth?: BoxProps['maxWidth'];
+  /**
    * If `true`, the card will be in selected state
    * Card will have a primary color border around it.
    *
    * @default false
    */
-
   isSelected?: boolean;
   /**
    * Makes the Card linkable by setting the `href` prop
@@ -185,6 +188,7 @@ const _Card: React.ForwardRefRenderFunction<BladeElementRef, CardProps> = (
     height,
     minHeight,
     minWidth,
+    maxWidth,
     onClick,
     isSelected = false,
     accessibilityLabel,
@@ -235,6 +239,7 @@ const _Card: React.ForwardRefRenderFunction<BladeElementRef, CardProps> = (
         height={height}
         minHeight={minHeight}
         minWidth={minWidth}
+        maxWidth={maxWidth}
         href={href}
         accessibilityLabel={accessibilityLabel}
         {...metaAttribute({ name: MetaConstants.Card, testID })}