+++ packages/blade/src/components/FileUpload/FileUpload.web.tsx
@@ -1,13 +1,20 @@
 import { useState, useCallback, useMemo, useRef, forwardRef } from 'react';
-import type { FileUploadProps, BladeFile, BladeFileList } from './types';
+import type {
+  FileUploadProps,
+  BladeFile,
+  BladeFileList,
+  FileUploadVariableSizeProps,
+} from './types';
 import { StyledFileUploadWrapper } from './StyledFileUploadWrapper';
 import {
   fileUploadColorTokens,
+  fileUploadHeightTokens,
   fileUploadLinkBorderTokens,
   getFileUploadInputHoverTokens,
 } from './fileUploadTokens';
 import { FileUploadItem } from './FileUploadItem';
 import { isFileAccepted } from './isFileAccepted';
+import { FileUploadItemIcon } from './FileUploadItemIcon';
 import BaseBox from '~components/Box/BaseBox';
 import { Box } from '~components/Box';
 import { SelectorLabel } from '~components/Form/Selector/SelectorLabel';
@@ -28,6 +35,7 @@ import { useMergeRefs } from '~utils/useMergeRefs';
 import { useControllableState } from '~utils/useControllable';
 import { getInnerMotionRef, getOuterMotionRef } from '~utils/getMotionRefs';
 import { makeAnalyticsAttribute } from '~utils/makeAnalyticsAttribute';
+import { throwBladeError } from '~utils/logger';
 
 const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadProps> = (
   {
@@ -59,6 +67,35 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
   },
   ref,
 ): React.ReactElement => {
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+  const { actionButtonText, dropAreaText, height, width } = rest as FileUploadVariableSizeProps;
+  const isSizeVariable = size === 'variable';
+  // 'actionButtonText', 'dropAreaText', 'height', 'width' are only valid when size is 'variable'
+  if (__DEV__) {
+    if (!isSizeVariable && (actionButtonText || dropAreaText)) {
+      const propName =
+        actionButtonText && dropAreaText
+          ? 'actionButtonText and dropAreaText'
+          : dropAreaText
+          ? 'dropAreaText'
+          : 'actionButtonText';
+
+      throwBladeError({
+        message: `${propName} can only be used when size is "variable"`,
+        moduleName: 'FileUpload',
+      });
+    }
+
+    if (!isSizeVariable && (height || width)) {
+      const propName = height && width ? 'height and width' : height ? 'height' : 'width';
+
+      throwBladeError({
+        message: `${propName} can only be used when size is "variable"`,
+        moduleName: 'FileUpload',
+      });
+    }
+  }
+
   const inputRef = useRef<HTMLInputElement | null>(null);
   const mergedRef = useMergeRefs(ref, inputRef);
   const { platform } = useTheme();
@@ -179,6 +216,20 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
     event.target.value = '';
   };
 
+  const getFileIconExtension = (acceptValue?: string): string => {
+    if (!acceptValue) return 'example.xyz';
+
+    const extensions = acceptValue
+      .split(',')
+      .map((ext) => ext.trim())
+      .filter((ext) => ext.startsWith('.'));
+
+    return extensions.length === 1 ? `example${extensions[0]}` : 'example.xyz';
+  };
+
+  const computedHeight = isSizeVariable ? height ?? '100%' : makeSize(fileUploadHeightTokens[size]);
+  const computedWidth = isSizeVariable ? width ?? '100%' : '100%';
+
   return (
     <BaseBox
       ref={getOuterMotionRef({ _motionMeta, ref })}
@@ -197,7 +248,7 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
       >
         {label ? (
           <FormLabel
-            size={size}
+            size={isSizeVariable ? 'large' : size}
             as="span"
             necessityIndicator={necessityIndicator}
             position={labelPosition}
@@ -213,7 +264,7 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
           inputProps={{}}
           style={{
             cursor: isDisabled ? 'not-allowed' : 'pointer',
-            width: '100%',
+            width: isOneFileSelectedWithSingleUpload && isLabelLeftPositioned ? undefined : '100%',
           }}
         >
           <BaseBox display="flex" flexDirection="column" width="100%">
@@ -225,6 +276,8 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
               flexDirection="row"
               justifyContent="center"
               alignItems="center"
+              height={computedHeight}
+              width={computedWidth}
               borderRadius="medium"
               borderWidth="thin"
               onDragOver={handleDragOver}
@@ -240,10 +293,17 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
                 display="flex"
                 justifyContent="center"
                 alignItems="center"
-                flexDirection={{ base: 'column', s: 'row' }}
+                flexDirection={{ base: 'column', s: `${isSizeVariable ? 'column' : 'row'}` }}
                 gap={makeSize(6)}
                 padding="spacing.3"
               >
+                {isSizeVariable && (
+                  <FileUploadItemIcon
+                    fileName={getFileIconExtension(accept)}
+                    uploadStatus="success"
+                  />
+                )}
+
                 <Text
                   color={
                     isDisabled
@@ -251,7 +311,9 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
                       : fileUploadColorTokens.text.default
                   }
                 >
-                  Drag files here or{' '}
+                  {isSizeVariable
+                    ? dropAreaText ?? 'Drag and drop your files here'
+                    : 'Drag files here or'}{' '}
                 </Text>
                 <SelectorInput
                   id={inputId}
@@ -296,7 +358,7 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
                           : fileUploadColorTokens.link.default
                       }
                     >
-                      Upload
+                      {actionButtonText ?? 'Upload'}
                     </Text>
                   </Box>
                 </Box>
@@ -340,12 +402,14 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
       {willRenderHintText && (
         <BaseBox
           marginLeft={makeSize(
-            label && isLabelLeftPositioned ? formHintLeftLabelMarginLeft[size] : 0,
+            label && isLabelLeftPositioned
+              ? formHintLeftLabelMarginLeft[isSizeVariable ? 'large' : size]
+              : 0,
           )}
         >
           <BaseBox display="flex" flexDirection="row" justifyContent="'space-between">
             <FormHint
-              size={size}
+              size={isSizeVariable ? 'large' : size}
               type={getHintType({
                 validationState: showError ? 'error' : validationState,
                 hasHelpText: Boolean(helpText),
@@ -363,7 +427,9 @@ const _FileUpload: React.ForwardRefRenderFunction<BladeElementRef, FileUploadPro
           <BaseBox
             key={file.id}
             marginLeft={makeSize(
-              label && isLabelLeftPositioned ? formHintLeftLabelMarginLeft[size] : 0,
+              label && isLabelLeftPositioned
+                ? formHintLeftLabelMarginLeft[isSizeVariable ? 'large' : size]
+                : 0,
             )}
             marginTop={index === 0 ? 'spacing.5' : 'spacing.3'}
           >